refactor(docs): spacer dx
