+++ apps/docs/content/components/spacer/usage.raw.jsx
@@ -1,0 +1,24 @@
+import {Spacer, Card} from "@nextui-org/react";
+
+export const CustomCard = () => (
+  <Card className="w-[200px] space-y-5 p-4" radius="2xl">
+    <div className="h-24 rounded-lg bg-default-300" />
+    <div className="space-y-3">
+      <div className="h-3 w-3/5 rounded-lg bg-default-200" />
+      <div className="h-3 w-4/5 rounded-lg bg-default-200" />
+      <div className="h-3 w-2/5 rounded-lg bg-default-300" />
+    </div>
+  </Card>
+);
+
+export default function App() {
+  return (
+    <div className="flex">
+      <CustomCard />
+      <Spacer x={4} />
+      <CustomCard />
+      <Spacer x={4} />
+      <CustomCard />
+    </div>
+  );
+}
